rename: validateInput 을 validateCardInfo 로 변경
