+++ src/domain/validateCardInfo.ts
